requirements.txt
pygame>=2.0.0
pillow>=9.0.0 